move add and remove axle unit functionality inside AxleUnitRow component
